Added ParticleGroup remove call to remove all particles in a ParticleSubGroup and clear method to empty the ParticleGroup
